changed all entity variable that were long to int to avoid ambiguity

+++ src/main/java/ch/uzh/ifi/hase/soprafs23/entity/game/Game.java
@@ -15,15 +15,15 @@ import java.util.List;
 public class Game implements Serializable {
 
     @Serial
-    private static final long serialVersionUID = 1L;
+    private static final int serialVersionUID = 1;
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "game_id")
-    private Long gameId;
+    private int gameId;
 
     @Column(nullable = false, unique = true)
-    private Long hostId;
+    private int hostId;
 
     @Column(nullable = false, unique = true)
     private int gamePin;
@@ -32,7 +32,7 @@ public class Game implements Serializable {
     private RoundLength roundLength;
 
     @Column(nullable = false)
-    private Long roundAmount;
+    private int roundAmount;
 
     @Column(nullable = false)
     private GameStatus status;
@@ -72,15 +72,15 @@ public class Game implements Serializable {
     @OneToMany(mappedBy = "game", cascade = CascadeType.ALL, orphanRemoval = true)
     private List<Round> rounds = new ArrayList<>();
 
-    public Long getGameId() {
+    public int getGameId() {
         return gameId;
     }
 
-    public Long getHostId() {
+    public int getHostId() {
         return hostId;
     }
 
-    public void setHostId(Long hostId) {
+    public void setHostId(int hostId) {
         this.hostId = hostId;
     }
 
@@ -100,11 +100,11 @@ public class Game implements Serializable {
         this.roundLength = roundLength;
     }
 
-    public Long getRounds() {
+    public int getRounds() {
         return roundAmount;
     }
 
-    public void setRounds(Long roundAmount) {
+    public void setRounds(int roundAmount) {
         this.roundAmount = roundAmount;
     }
