+++ src/main/java/ch/uzh/ifi/hase/soprafs23/rest/dto/game/WinnerDTO.java
@@ -5,7 +5,7 @@ import ch.uzh.ifi.hase.soprafs23.entity.User;
 public class WinnerDTO {
 
     private User user;
-    private Long score;
+    private int score;
 
 
     public User getUser() {
@@ -16,11 +16,11 @@ public class WinnerDTO {
         this.user = user;
     }
 
-    public Long getScore() {
+    public int getScore() {
         return score;
     }
 
-    public void setScore(Long score) {
+    public void setScore(int score) {
         this.score = score;
     }
 }
